Simplify before rewriting again
