+++ src/global_metadata.rs
@@ -706,49 +706,6 @@ struct Il2CppSectionMetadata {
     count: u32,
 }
 
-/// The width (in bytes) that each variable-width index kind is serialized at
-/// in this file. Introduced in v39: see the comment on [`Il2CppSectionMetadata::count`].
-#[cfg(feature = "il2cpp_v39")]
-#[derive(Debug, Clone, Copy)]
-struct IndexSizes {
-    /// Width (in bytes) of `TypeIndex` fields, sized against the *runtime*
-    /// metadata's `types` array (not anything in this file), since that's
-    /// the table `TypeIndex` actually indexes into.
-    type_index: usize,
-    type_definition_index: usize,
-    generic_container_index: usize,
-    parameter_index: usize,
-}
-
-#[cfg(feature = "il2cpp_v39")]
-fn index_size(num_elements: u32) -> usize {
-    if num_elements <= u8::MAX as u32 {
-        1
-    } else if num_elements <= u16::MAX as u32 {
-        2
-    } else {
-        4
-    }
-}
-
-/// Reads a variable-width index: an unsigned integer of `width` bytes whose
-/// max value (the "no value" sentinel) is normalized to `u32::MAX`, matching
-/// the fixed-width indices elsewhere in this file (see `IndexType::is_valid`).
-#[cfg(feature = "il2cpp_v39")]
-fn read_var_index(cursor: &mut Cursor<&[u8]>, width: usize) -> std::io::Result<u32> {
-    Ok(match width {
-        1 => {
-            let v = cursor.read_u8()?;
-            if v == u8::MAX { u32::MAX } else { v as u32 }
-        }
-        2 => {
-            let v = cursor.read_u16::<LittleEndian>()?;
-            if v == u16::MAX { u32::MAX } else { v as u32 }
-        }
-        _ => cursor.read_u32::<LittleEndian>()?,
-    })
-}
-
 trait ReadMetadataTable<'a>
 where
     Self: std::marker::Sized,
@@ -759,6 +716,7 @@ where
 macro_rules! metadata {
     ($($(#[$($attrss:tt)*])* $name:ident: $ty:ty,)*) => {
         #[derive(Debug, BinaryDeserialize)]
+        #[allow(unused)]
         struct Il2CppGlobalMetadataHeader {
             sanity: u32,
             version: u32,
@@ -776,7 +734,6 @@ macro_rules! metadata {
         }
 
         impl<'a> GlobalMetadata<'a> {
-            #[cfg(feature = "il2cpp_v31")]
             fn deserialize(
                 data: &'a [u8],
                 header: Il2CppGlobalMetadataHeader,
@@ -1039,370 +996,6 @@ metadata! {
     exported_type_definitions: ExportedTypeDefinitionTable,
 }
 
-// v39 entry readers for tables containing variable-width indices (see
-// `IndexSizes`), generated by `var_width_table!` from a compact per-field
-// spec instead of being hand-written:
-//   `field: fixed(Type)`       - read via `Type`'s normal fixed-width decode
-//   `field: var(size)`         - a bare variable-width `TypeIndex`
-//   `field: var(size, Ctor)`   - a variable-width index wrapped with `Ctor`
-//     (always `SomeIndexType::new`; `size` names the `IndexSizes` field
-//     that determines its width)
-// Field order per table must match the file's real byte layout
-// (cross-checked against libil2cpp's own `vm/MetadataDeserialization.cpp`),
-// which is not always the same as the Rust struct's declaration order.
-#[cfg(feature = "il2cpp_v39")]
-macro_rules! var_width_field {
-    ($cur:expr, $sizes:expr, fixed($ty:path)) => {
-        <$ty as BinaryDeserialize>::deserialize::<LittleEndian, _>(&mut *$cur)?
-    };
-    ($cur:expr, $sizes:expr, var($size:ident)) => {
-        read_var_index($cur, $sizes.$size)?
-    };
-    ($cur:expr, $sizes:expr, var($size:ident, $ctor:path)) => {
-        $ctor(read_var_index($cur, $sizes.$size)?)
-    };
-}
-
-/// The number of bytes [`var_width_field!`] consumes for the same field spec.
-#[cfg(feature = "il2cpp_v39")]
-macro_rules! var_width_field_size {
-    ($sizes:expr, fixed($ty:path)) => {
-        <$ty as BinaryDeserialize>::SIZE
-    };
-    ($sizes:expr, var($size:ident)) => {
-        $sizes.$size
-    };
-    ($sizes:expr, var($size:ident, $ctor:path)) => {
-        $sizes.$size
-    };
-}
-
-#[cfg(feature = "il2cpp_v39")]
-macro_rules! var_width_table {
-    ($read_fn:ident, $stride_fn:ident -> $ty:ident {
-        $($field:ident : $kind:ident ( $($args:tt)* )),* $(,)?
-    }) => {
-        pub(super) fn $read_fn(
-            cur: &mut Cursor<&[u8]>,
-            sizes: IndexSizes,
-        ) -> std::io::Result<$ty> {
-            Ok($ty {
-                $($field: var_width_field!(cur, sizes, $kind($($args)*)),)*
-            })
-        }
-
-        pub(super) fn $stride_fn(sizes: IndexSizes) -> usize {
-            0 $(+ var_width_field_size!(sizes, $kind($($args)*)))*
-        }
-    };
-}
-
-#[cfg(feature = "il2cpp_v39")]
-mod var_width_readers {
-    use super::*;
-
-    var_width_table!(event_definition, event_definition_stride -> Il2CppEventDefinition {
-        name_index: fixed(StringIndex),
-        type_index: var(type_index),
-        add: fixed(MethodIndex),
-        remove: fixed(MethodIndex),
-        raise: fixed(MethodIndex),
-        token: fixed(Token),
-    });
-
-    var_width_table!(method_definition, method_definition_stride -> Il2CppMethodDefinition {
-        name_index: fixed(StringIndex),
-        declaring_type: var(type_definition_index, TypeDefinitionIndex::new),
-        return_type: var(type_index),
-        return_parameter_token: fixed(Token),
-        parameter_start: var(parameter_index, ParameterIndex::new),
-        generic_container_index: var(generic_container_index, GenericContainerIndex::new),
-        token: fixed(Token),
-        flags: fixed(u16),
-        iflags: fixed(u16),
-        slot: fixed(u16),
-        parameter_count: fixed(u16),
-    });
-
-    var_width_table!(parameter_default_value, parameter_default_value_stride -> Il2CppParameterDefaultValue {
-        parameter_index: var(parameter_index, ParameterIndex::new),
-        type_index: var(type_index),
-        data_index: fixed(FieldAndParameterDefaultValueIndex),
-    });
-
-    var_width_table!(field_default_value, field_default_value_stride -> Il2CppFieldDefaultValue {
-        field_index: fixed(FieldIndex),
-        type_index: var(type_index),
-        data_index: fixed(FieldAndParameterDefaultValueIndex),
-    });
-
-    var_width_table!(field_marshaled_size, field_marshaled_size_stride -> Il2CppFieldMarshaledSize {
-        field_index: fixed(FieldIndex),
-        type_index: var(type_index),
-        size: fixed(u32),
-    });
-
-    var_width_table!(parameter_definition, parameter_definition_stride -> Il2CppParameterDefinition {
-        name_index: fixed(StringIndex),
-        token: fixed(Token),
-        type_index: var(type_index),
-    });
-
-    var_width_table!(field_definition, field_definition_stride -> Il2CppFieldDefinition {
-        name_index: fixed(StringIndex),
-        type_index: var(type_index),
-        token: fixed(Token),
-    });
-
-    var_width_table!(interface_offset_pair, interface_offset_pair_stride -> Il2CppInterfaceOffsetPair {
-        interface_type_index: var(type_index),
-        offset: fixed(u32),
-    });
-
-    var_width_table!(type_definition, type_definition_stride -> Il2CppTypeDefinition {
-        name_index: fixed(StringIndex),
-        namespace_index: fixed(StringIndex),
-        byval_type_index: var(type_index),
-        declaring_type_index: var(type_index),
-        parent_index: var(type_index),
-        generic_container_index: var(generic_container_index, GenericContainerIndex::new),
-        flags: fixed(u32),
-        field_start: fixed(FieldIndex),
-        method_start: fixed(MethodIndex),
-        event_start: fixed(EventIndex),
-        property_start: fixed(PropertyIndex),
-        nested_types_start: fixed(NestedTypeIndex),
-        interfaces_start: fixed(InterfaceIndex),
-        vtable_start: fixed(VTableMethodIndex),
-        interface_offsets_start: fixed(InterfaceOffsetIndex),
-        method_count: fixed(u16),
-        property_count: fixed(u16),
-        field_count: fixed(u16),
-        event_count: fixed(u16),
-        nested_type_count: fixed(u16),
-        vtable_count: fixed(u16),
-        interfaces_count: fixed(u16),
-        interface_offsets_count: fixed(u16),
-        bitfield: fixed(u32),
-        token: fixed(Token),
-    });
-
-    var_width_table!(field_ref, field_ref_stride -> Il2CppFieldRef {
-        type_index: var(type_index),
-        field_index: fixed(u32),
-    });
-
-    var_width_table!(image_definition, image_definition_stride -> Il2CppImageDefinition {
-        name_index: fixed(StringIndex),
-        assembly_index: fixed(AssemblyIndex),
-        type_start: var(type_definition_index, TypeDefinitionIndex::new),
-        type_count: fixed(u32),
-        exported_type_start: var(type_definition_index, TypeDefinitionIndex::new),
-        exported_type_count: fixed(u32),
-        entry_point_index: fixed(MethodIndex),
-        token: fixed(Token),
-        custom_attribute_start: fixed(AttributeDataRangeIndex),
-        custom_attribute_count: fixed(u32),
-    });
-
-    var_width_table!(generic_parameter, generic_parameter_stride -> Il2CppGenericParameter {
-        owner_index: var(generic_container_index, GenericContainerIndex::new),
-        name_index: fixed(StringIndex),
-        constraints_start: fixed(GenericParameterConstraintIndex),
-        constraints_count: fixed(u16),
-        num: fixed(u16),
-        flags: fixed(u16),
-    });
-
-    var_width_table!(windows_runtime_type_name_pair, windows_runtime_type_name_pair_stride -> Il2CppWindowsRuntimeTypeNamePair {
-        name_index: fixed(StringIndex),
-        type_index: var(type_index),
-    });
-}
-
-#[cfg(feature = "il2cpp_v39")]
-impl<'a> GlobalMetadata<'a> {
-    fn deserialize(
-        data: &'a [u8],
-        header: Il2CppGlobalMetadataHeader,
-        types_count: u32,
-    ) -> Result<GlobalMetadata<'a>, MetadataDeserializeError> {
-        use var_width_readers as r;
-
-        let sizes = IndexSizes {
-            type_index: index_size(types_count),
-            type_definition_index: index_size(header.type_definitions.count),
-            generic_container_index: index_size(header.generic_containers.count),
-            parameter_index: index_size(header.parameters.count),
-        };
-
-        let mut cursor = Cursor::new(data);
-
-        // Sections whose element layout doesn't depend on `sizes`: read with
-        // the same fixed-size machinery v31 uses unconditionally.
-        macro_rules! fixed_section {
-            ($name:ident) => {{
-                let size = header.$name.len as usize;
-                if size > 0 {
-                    cursor.set_position(header.$name.offset as u64);
-                    ReadMetadataTable::read(&mut cursor, size)?
-                } else {
-                    Default::default()
-                }
-            }};
-        }
-
-        // Tables of a struct type containing variable-width indices: read
-        // entry-by-entry with the given reader, using its matching
-        // `*_stride` function (both generated by `var_width_table!`) to
-        // compute how many entries the section holds.
-        macro_rules! var_section {
-            ($name:ident, $read_one:path, $stride_fn:path, $table:ident) => {{
-                let size = header.$name.len as usize;
-                if size > 0 {
-                    cursor.set_position(header.$name.offset as u64);
-                    let count = size / $stride_fn(sizes);
-                    let mut vec = Vec::with_capacity(count);
-                    for _ in 0..count {
-                        vec.push($read_one(&mut cursor, sizes)?);
-                    }
-                    $table { table: vec }
-                } else {
-                    Default::default()
-                }
-            }};
-        }
-
-        // Raw `TypeIndex` arrays (not wrapped in a struct): same deal, but
-        // each element is a bare variable-width index.
-        macro_rules! var_index_section {
-            ($name:ident, $table:ident) => {{
-                let size = header.$name.len as usize;
-                if size > 0 {
-                    cursor.set_position(header.$name.offset as u64);
-                    let count = size / sizes.type_index;
-                    let mut vec = Vec::with_capacity(count);
-                    for _ in 0..count {
-                        vec.push(read_var_index(&mut cursor, sizes.type_index)?);
-                    }
-                    $table { table: vec }
-                } else {
-                    Default::default()
-                }
-            }};
-        }
-
-        Ok(GlobalMetadata {
-            string_literal: fixed_section!(string_literal),
-            string_literal_data: fixed_section!(string_literal_data),
-            string: fixed_section!(string),
-            events: var_section!(
-                events,
-                r::event_definition,
-                r::event_definition_stride,
-                EventTable
-            ),
-            properties: fixed_section!(properties),
-            methods: var_section!(
-                methods,
-                r::method_definition,
-                r::method_definition_stride,
-                MethodTable
-            ),
-            parameter_default_values: var_section!(
-                parameter_default_values,
-                r::parameter_default_value,
-                r::parameter_default_value_stride,
-                ParameterDefaultValueTable
-            ),
-            field_default_values: var_section!(
-                field_default_values,
-                r::field_default_value,
-                r::field_default_value_stride,
-                FieldDefaultValueTable
-            ),
-            field_and_parameter_default_value_data: fixed_section!(
-                field_and_parameter_default_value_data
-            ),
-            field_marshaled_sizes: var_section!(
-                field_marshaled_sizes,
-                r::field_marshaled_size,
-                r::field_marshaled_size_stride,
-                FieldMarshaledSizeTable
-            ),
-            parameters: var_section!(
-                parameters,
-                r::parameter_definition,
-                r::parameter_definition_stride,
-                ParameterTable
-            ),
-            fields: var_section!(
-                fields,
-                r::field_definition,
-                r::field_definition_stride,
-                FieldTable
-            ),
-            generic_parameters: var_section!(
-                generic_parameters,
-                r::generic_parameter,
-                r::generic_parameter_stride,
-                GenericParameterTable
-            ),
-            generic_parameter_constraints: var_index_section!(
-                generic_parameter_constraints,
-                GenericParameterConstraintTable
-            ),
-            generic_containers: fixed_section!(generic_containers),
-            nested_types: fixed_section!(nested_types),
-            interfaces: var_index_section!(interfaces, InterfaceTable),
-            vtable_methods: fixed_section!(vtable_methods),
-            interface_offsets: var_section!(
-                interface_offsets,
-                r::interface_offset_pair,
-                r::interface_offset_pair_stride,
-                InterfaceOffsetTable
-            ),
-            type_definitions: var_section!(
-                type_definitions,
-                r::type_definition,
-                r::type_definition_stride,
-                TypeDefinitionTable
-            ),
-            images: var_section!(
-                images,
-                r::image_definition,
-                r::image_definition_stride,
-                ImageTable
-            ),
-            assemblies: fixed_section!(assemblies),
-            field_refs: var_section!(
-                field_refs,
-                r::field_ref,
-                r::field_ref_stride,
-                FieldRefTable
-            ),
-            referenced_assemblies: fixed_section!(referenced_assemblies),
-            attribute_data: fixed_section!(attribute_data),
-            attribute_data_range: fixed_section!(attribute_data_range),
-            unresolved_indirect_call_parameter_types: var_index_section!(
-                unresolved_indirect_call_parameter_types,
-                UnresolvedIndirectCallParameterTypeTable
-            ),
-            unresolved_indirect_call_parameter_ranges: fixed_section!(
-                unresolved_indirect_call_parameter_ranges
-            ),
-            windows_runtime_type_names: var_section!(
-                windows_runtime_type_names,
-                r::windows_runtime_type_name_pair,
-                r::windows_runtime_type_name_pair_stride,
-                WindowsRuntimeTypeNameTable
-            ),
-            windows_runtime_strings: fixed_section!(windows_runtime_strings),
-            exported_type_definitions: fixed_section!(exported_type_definitions),
-        })
-    }
-}
-
 #[derive(Error, Debug)]
 pub enum MetadataDeserializeError {
     #[error("binary deserialization error")]
@@ -1515,8 +1108,6 @@ mod tests {
     /// A row count large enough that `index_size` always picks 4 bytes.
     #[cfg(feature = "il2cpp_v39")]
     const VAR_INDEX_COUNT: u32 = u16::MAX as u32 + 1;
-    #[cfg(feature = "il2cpp_v31")]
-    const VAR_INDEX_COUNT: u32 = 0;
 
     /// Deserializes a buffer built by [`build_metadata`], hiding v39's extra
     /// `types_count` parameter (also forced to 4-byte width; see `VAR_INDEX_COUNT`).
@@ -1524,7 +1115,7 @@ mod tests {
         #[cfg(feature = "il2cpp_v31")]
         return deserialize(data);
         #[cfg(feature = "il2cpp_v39")]
-        return deserialize(data, VAR_INDEX_COUNT);
+        return deserialize(data);
     }
 
     #[test]
